requirements/development.txt
-r base.txt

django-extensions==1.5.6
